Immediately write pulse card graphs to disk
Previously images in pulse emails were serialized to bytes and the
bytes were kept in map stored in an atom. After all of the graphs were
serialized, they were written to disk and the file handle was passed
to the email library.

This commit immediately writes the file to disk and returns a
URL. This will no longer hold onto the memory need for all of the
cards and can be GC'd right away. This also avoids writing static
images like the no results image to disk evertime, rather than just
reading it out of the JAR.